Rename common.py to coordinator.py
