+++ homeassistant/components/withings/coordinator.py
