+++ homeassistant/components/withings/sensor.py
@@ -23,7 +23,6 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import WithingsDataUpdateCoordinator
 from .const import (
     DOMAIN,
     SCORE_POINTS,
@@ -33,6 +32,7 @@ from .const import (
     UOM_MMHG,
     Measurement,
 )
+from .coordinator import WithingsDataUpdateCoordinator
 from .entity import WithingsEntity, WithingsEntityDescription
