+++ tests/components/withings/conftest.py
@@ -142,10 +142,10 @@ def disable_webhook_delay():
 
     mock = AsyncMock()
     with patch(
-        "homeassistant.components.withings.common.SUBSCRIBE_DELAY",
+        "homeassistant.components.withings.coordinator.SUBSCRIBE_DELAY",
         timedelta(seconds=0),
     ), patch(
-        "homeassistant.components.withings.common.UNSUBSCRIBE_DELAY",
+        "homeassistant.components.withings.coordinator.UNSUBSCRIBE_DELAY",
         timedelta(seconds=0),
     ):
         yield mock
